refactor(client): move single-client navigation/backup methods out of base.py
- Move `back_up_branch_layout_file` (only used by `discover`) into
  `DiscoverMacheteClient`; `import shutil` follows it.
- Move the seven navigation primitives used solely by the `go-show`
  subcommand (`first_branch_for`, `last_branch_for`, `next_branch_for`,
  `prev_branch_for`, `first_root_branch`, `last_root_branch`,
  `get_or_pick_child_of`) into `GoShowMacheteClient`.
- Rename `MacheteClient.__raise_no_branches_error` to
  `_raise_no_branches_error` so subclasses can call it without name
  mangling - it stays callable from `MacheteClient` itself and from
  `GoShowMacheteClient.first_root_branch`/`last_root_branch`.
